# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dtorkon API",
    "description": "FastAPI backend for the dtorkon mini-blog. The service handles admin-only auth, public/admin posts, site profile editing, analytics, SQLite persistence, searchable publishing, and backend-mediated uploads to Yandex Object Storage over the S3 API.",
    "version": "0.1.0-fastapi"
  },
  "servers": [
    {
      "url": "https://{host}",
      "variables": {
        "host": {
          "default": "example.com"
        }
      }
    }
  ],
  "paths": {
...
    "/api/posts": {
      "get": {
        "tags": [
          "Public Posts"
        ],
        "summary": "Get a paginated list of published posts",
        "operationId": "list_public_posts_api_posts_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 12,
              "title": "Page Size"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicPostListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/{slug}": {
      "get": {
        "tags": [
          "Public Posts"
        ],
        "summary": "Get a published post by slug",
        "operationId": "get_public_post_api_posts__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicPostDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/analytics": {
      "get": {
        "tags": [
          "Admin Analytics"
        ],
        "summary": "Get analytics for the author workspace",
        "operationId": "get_admin_analytics_api_admin_analytics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAnalyticsRead"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/api/admin/posts": {
      "get": {
        "tags": [
          "Admin Posts"
        ],
        "summary": "Get posts for the author workspace",
        "operationId": "list_admin_posts_api_admin_posts_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PostStatus"
                },
                {
                  "type": "string"
                }
              ],
              "default": "all",
              "title": "Status"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPostListResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin Posts"
        ],
        "summary": "Create a new post",
        "operationId": "create_post_api_admin_posts_post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPostDetail"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/admin/posts/{post_id}": {
      "get": {
        "tags": [
          "Admin Posts"
        ],
        "summary": "Get one post for editing",
        "operationId": "get_admin_post_api_admin_posts__post_id__get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPostDetail"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Admin Posts"
        ],
        "summary": "Update an existing post",
        "operationId": "update_post_api_admin_posts__post_id__patch",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPostDetail"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin Posts"
        ],
        "summary": "Delete a post",
        "operationId": "delete_post_api_admin_posts__post_id__delete",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminAnalyticsRead": {
        "properties": {
          "totalPosts": {
            "type": "integer",
            "title": "Totalposts"
          },
          "publishedPosts": {
            "type": "integer",
            "title": "Publishedposts"
          },
          "draftPosts": {
            "type": "integer",
            "title": "Draftposts"
          },
          "totalAssets": {
            "type": "integer",
            "title": "Totalassets"
          },
          "readyAssets": {
            "type": "integer",
            "title": "Readyassets"
          },
          "totalWords": {
            "type": "integer",
            "title": "Totalwords"
          },
          "totalAttachments": {
            "type": "integer",
            "title": "Totalattachments"
          },
          "transcriptReady": {
            "type": "integer",
            "title": "Transcriptready"
          },
          "transcriptProcessing": {
            "type": "integer",
            "title": "Transcriptprocessing"
          },
          "transcriptFailed": {
            "type": "integer",
            "title": "Transcriptfailed"
          },
          "publicationActivity": {
            "items": {
              "$ref": "#/components/schemas/AnalyticsTimelinePoint"
            },
            "type": "array",
            "title": "Publicationactivity"
          },
          "uploadActivity": {
            "items": {
              "$ref": "#/components/schemas/AnalyticsTimelinePoint"
            },
            "type": "array",
            "title": "Uploadactivity"
          },
          "assetBreakdown": {
            "items": {
              "$ref": "#/components/schemas/AnalyticsBreakdownItem"
            },
            "type": "array",
            "title": "Assetbreakdown"
          },
          "totalErrors": {
            "type": "integer",
            "title": "Totalerrors"
          },
          "lastErrorAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lasterrorat"
          },
          "recentErrors": {
            "items": {
              "$ref": "#/components/schemas/AdminErrorEventRead"
            },
            "type": "array",
            "title": "Recenterrors"
          }
        },
        "type": "object",
        "required": [
          "totalPosts",
          "publishedPosts",
          "draftPosts",
          "totalAssets",
          "readyAssets",
          "totalWords",
          "totalAttachments",
          "transcriptReady",
          "transcriptProcessing",
          "transcriptFailed",
          "publicationActivity",
          "uploadActivity",
          "assetBreakdown",
          "totalErrors"
        ],
        "title": "AdminAnalyticsRead"
      },
...
      "AdminPostDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          },
          "bodyMarkdown": {
            "type": "string",
            "title": "Bodymarkdown"
          },
          "status": {
            "$ref": "#/components/schemas/PostStatus"
          },
          "coverAssetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverassetid"
          },
          "coverAsset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "inlineAssets": {
            "items": {
              "$ref": "#/components/schemas/AssetRead"
            },
            "type": "array",
            "title": "Inlineassets"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentRead"
            },
            "type": "array",
            "title": "Attachments"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          },
          "publishedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publishedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "excerpt",
          "bodyMarkdown",
          "status",
          "attachments",
          "createdAt",
          "updatedAt"
        ],
        "title": "AdminPostDetail"
      },
      "AdminPostListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminPostSummary"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "AdminPostListResponse"
      },
...
      "AttachmentRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "assetId": {
            "type": "string",
            "title": "Assetid"
          },
          "kind": {
            "$ref": "#/components/schemas/AttachmentKind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "sortOrder": {
            "type": "integer",
            "title": "Sortorder"
          },
          "asset": {
            "$ref": "#/components/schemas/AssetRead"
          }
        },
        "type": "object",
        "required": [
          "id",
          "assetId",
          "kind",
          "title",
          "sortOrder",
          "asset"
        ],
        "title": "AttachmentRead"
      },
...
      "CreatePostRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt",
            "default": ""
          },
          "bodyMarkdown": {
            "type": "string",
            "title": "Bodymarkdown"
          },
          "status": {
            "$ref": "#/components/schemas/PostStatus"
          },
          "coverAssetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverassetid"
          },
          "inlineAssetIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Inlineassetids"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentInput"
            },
            "type": "array",
            "title": "Attachments"
          }
        },
        "type": "object",
        "required": [
          "title",
          "slug",
          "bodyMarkdown",
          "status"
        ],
        "title": "CreatePostRequest"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorPayload"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Pagination": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "pageSize": {
            "type": "integer",
            "title": "Pagesize"
          },
          "totalItems": {
            "type": "integer",
            "title": "Totalitems"
          },
          "totalPages": {
            "type": "integer",
            "title": "Totalpages"
          }
        },
        "type": "object",
        "required": [
          "page",
          "pageSize",
          "totalItems",
          "totalPages"
        ],
        "title": "Pagination"
      },
      "PostStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published"
        ],
        "title": "PostStatus"
      },
...
      "PublicPostDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          },
          "bodyMarkdown": {
            "type": "string",
            "title": "Bodymarkdown"
          },
          "status": {
            "$ref": "#/components/schemas/PostStatus"
          },
          "coverAsset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentRead"
            },
            "type": "array",
            "title": "Attachments"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          },
          "publishedAt": {
            "type": "string",
            "title": "Publishedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "excerpt",
          "bodyMarkdown",
          "status",
          "attachments",
          "createdAt",
          "updatedAt",
          "publishedAt"
        ],
        "title": "PublicPostDetail"
      },
      "PublicPostListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          },
          "coverAsset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "publishedAt": {
            "type": "string",
            "title": "Publishedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "excerpt",
          "publishedAt"
        ],
        "title": "PublicPostListItem"
      },
      "PublicPostListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PublicPostListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "type": "object",
        "required": [
          "items",
          "pagination"
        ],
        "title": "PublicPostListResponse"
      },
...
      "UpdatePostRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt"
          },
          "bodyMarkdown": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bodymarkdown"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PostStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "coverAssetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverassetid"
          },
          "inlineAssetIds": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inlineassetids"
          },
          "attachments": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AttachmentInput"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachments"
          }
        },
        "type": "object",
        "title": "UpdatePostRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dtorkon_session"
      }
    }
  },
...
}